refactor: extract handler for testability, concurrent liveness pings, bridge reaping (#45)
- Extract createFetchHandler into handler.ts so tests use real routing logic
- Make liveness pings concurrent with Promise.allSettled (prevents overlap)
- Reap bridges inactive for 5x timeout to prevent unbounded memory growth
- Fail loud on missing GATEWAY_SECRET instead of silently rejecting all registrations
- DRY up auth+parse into parseAuthenticatedBody helper
- Unify SIGTERM/SIGINT into single shutdown function
